requirements.txt
Django==5.0.8
celery @ https://github.com/celery/celery/zipball/main#sha256=1314ac3d2c5cad2fcbfd6cb44359171a35cbd59a6b52882471e9c920f2829362
freezegun==1.5.1

## The following requirements were added by pip freeze:
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astroid==3.2.4
attrs==24.2.0
autobahn==24.4.2
Automat==22.10.0
billiard==4.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.0
dill==0.3.8
distro==1.9.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-filter==24.3
django-grip==3.5.0
djangorestframework==3.15.2
djangorestframework-jsonapi==7.0.2
docstring_parser==0.16
frozenlist==1.4.1
gripcontrol==4.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
instructor==1.3.7
isort==5.13.2
jiter==0.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.0
langchain==0.2.12
langchain-core==0.2.29
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langsmith==0.1.98
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
openai==1.40.2
orjson==3.10.7
packaging==24.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
pubcontrol==3.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pylint==3.2.6
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.7.1
service-identity==24.1.0
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlparse==0.5.1
tenacity==8.5.0
tiktoken==0.7.0
tomlkit==0.13.0
tqdm==4.66.5
Twisted==24.3.0
txaio==23.1.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
yarl==1.9.4
zope.interface==7.0.1